Added moodle to only live site

+++ app/Console/Commands/VisitAgreement.php
@@ -69,34 +69,36 @@ class VisitAgreement extends Command
                 $user->added_to_facility = substr($r->facility_join, 0, 10).' '.substr($r->facility_join, 11, 8);
                 $user->save();
 
-                //Assigns role in moodle database
-                if($rating_old != $r->rating) {
-                    $old_role = DB::table('mdl_role_assignments')->where('userid', $user->id)->where('roleid', '!=', 14)->where('roleid', '!=', 15)->where('roleid', '!=', 16)->where('roleid', '!=', 17)->first();
-                    $old_role->delete();
+                if(Config::get('app.moodle') == 1) {
+                    //Assigns role in moodle database
+                    if($rating_old != $r->rating) {
+                        $old_role = DB::table('mdl_role_assignments')->where('userid', $user->id)->where('roleid', '!=', 14)->where('roleid', '!=', 15)->where('roleid', '!=', 16)->where('roleid', '!=', 17)->first();
+                        $old_role->delete();
 
-                    if($user->rating_id == 1) {
-                        $mdl_rating = 18;
-                    } elseif($user->rating_id == 2) {
-                        $mdl_rating = 9;
-                    } elseif($user->rating_id == 3) {
-                        $mdl_rating = 10;
-                    } elseif($user->rating_id == 4) {
-                        $mdl_rating = 11;
-                    } elseif($user->rating_id == 5) {
-                        $mdl_rating = 12;
-                    } elseif($user->rating_id == 7 || $user->rating_id == 11 || $user->rating_id == 12) {
-                        $mdl_rating = 13;
-                    } elseif($user->rating_id == 8 || $user->rating_id == 10) {
-                        $mdl_rating = 14;
-                    } else {
-                        $mdl_rating = 0;
+                        if($user->rating_id == 1) {
+                            $mdl_rating = 18;
+                        } elseif($user->rating_id == 2) {
+                            $mdl_rating = 9;
+                        } elseif($user->rating_id == 3) {
+                            $mdl_rating = 10;
+                        } elseif($user->rating_id == 4) {
+                            $mdl_rating = 11;
+                        } elseif($user->rating_id == 5) {
+                            $mdl_rating = 12;
+                        } elseif($user->rating_id == 7 || $user->rating_id == 11 || $user->rating_id == 12) {
+                            $mdl_rating = 13;
+                        } elseif($user->rating_id == 8 || $user->rating_id == 10) {
+                            $mdl_rating = 14;
+                        } else {
+                            $mdl_rating = 0;
+                        }
+
+                        DB::table('mdl_role_assignments')->insert([
+                            'roleid' => $mdl_rating,
+                            'contextid' => 26,
+                            'userid' => $user->id
+                        ]);
                     }
-
-                    DB::table('mdl_role_assignments')->insert([
-                        'roleid' => $mdl_rating,
-                        'contextid' => 26,
-                        'userid' => $user->id
-                    ]);
                 }
             } else {
                 $visitrej = VisitRej::where('cid', $r->cid)->first();
@@ -126,41 +128,43 @@ class VisitAgreement extends Command
                     $user->added_to_facility = substr($r->facility_join, 0, 10).' '.substr($r->facility_join, 11, 8);
                     $user->save();
 
-                    //Adds user to moodle database
-                    DB::table('mdl_user')->insert([
-                         'id' => $user->id,
-                         'confirmed' => 1,
-                         'mnethostid' => 1,
-                         'username' => $user->id,
-                         'firstname' => $user->fname,
-                         'lastname' => $user->lname,
-                         'email' => $user->email
-                     ]);
+                    if(Config::get('app.moodle') == 1) {
+                        //Adds user to moodle database
+                        DB::table('mdl_user')->insert([
+                             'id' => $user->id,
+                             'confirmed' => 1,
+                             'mnethostid' => 1,
+                             'username' => $user->id,
+                             'firstname' => $user->fname,
+                             'lastname' => $user->lname,
+                             'email' => $user->email
+                         ]);
 
-                    //Assigns role in moodle database
-                    if($user->rating_id == 1) {
-                        $mdl_rating = 18;
-                    } elseif($user->rating_id == 2) {
-                        $mdl_rating = 9;
-                    } elseif($user->rating_id == 3) {
-                        $mdl_rating = 10;
-                    } elseif($user->rating_id == 4) {
-                        $mdl_rating = 11;
-                    } elseif($user->rating_id == 5) {
-                        $mdl_rating = 12;
-                    } elseif($user->rating_id == 7 || $user->rating_id == 11 || $user->rating_id == 12) {
-                        $mdl_rating = 13;
-                    } elseif($user->rating_id == 8 || $user->rating_id == 10) {
-                        $mdl_rating = 14;
-                    } else {
-                        $mdl_rating = 0;
+                        //Assigns role in moodle database
+                        if($user->rating_id == 1) {
+                            $mdl_rating = 18;
+                        } elseif($user->rating_id == 2) {
+                            $mdl_rating = 9;
+                        } elseif($user->rating_id == 3) {
+                            $mdl_rating = 10;
+                        } elseif($user->rating_id == 4) {
+                            $mdl_rating = 11;
+                        } elseif($user->rating_id == 5) {
+                            $mdl_rating = 12;
+                        } elseif($user->rating_id == 7 || $user->rating_id == 11 || $user->rating_id == 12) {
+                            $mdl_rating = 13;
+                        } elseif($user->rating_id == 8 || $user->rating_id == 10) {
+                            $mdl_rating = 14;
+                        } else {
+                            $mdl_rating = 0;
+                        }
+
+                        DB::table('mdl_role_assignments')->insert([
+                            'roleid' => $mdl_rating,
+                            'contextid' => 26,
+                            'userid' => $user->id
+                        ]);
                     }
-
-                    DB::table('mdl_role_assignments')->insert([
-                        'roleid' => $mdl_rating,
-                        'contextid' => 26,
-                        'userid' => $user->id
-                    ]);
 
                     //Assigns controller initials
                     $user = User::find($r->cid);
@@ -319,45 +323,36 @@ class VisitAgreement extends Command
                 }
                 $user->save();
 
-                //Adds user to moodle database
-                DB::table('mdl_user')->insert([
-                     'id' => $user->id,
-                     'confirmed' => 1,
-                     'mnethostid' => 1,
-                     'username' => $user->id,
-                     'firstname' => $user->fname,
-                     'lastname' => $user->lname,
-                     'email' => $user->email
-                 ]);
+                if(Config::get('app.moodle') == 1) {
+                    //Assigns role in moodle database
+                    if($rating_old != $r->rating) {
+                        $old_role = DB::table('mdl_role_assignments')->where('userid', $user->id)->where('roleid', '!=', 14)->where('roleid', '!=', 15)->where('roleid', '!=', 16)->where('roleid', '!=', 17)->first();
+                        $old_role->delete();
 
-                //Assigns role in moodle database
-                if($rating_old != $r->rating) {
-                    $old_role = DB::table('mdl_role_assignments')->where('userid', $user->id)->where('roleid', '!=', 14)->where('roleid', '!=', 15)->where('roleid', '!=', 16)->where('roleid', '!=', 17)->first();
-                    $old_role->delete();
+                        if($user->rating_id == 1) {
+                            $mdl_rating = 18;
+                        } elseif($user->rating_id == 2) {
+                            $mdl_rating = 9;
+                        } elseif($user->rating_id == 3) {
+                            $mdl_rating = 10;
+                        } elseif($user->rating_id == 4) {
+                            $mdl_rating = 11;
+                        } elseif($user->rating_id == 5) {
+                            $mdl_rating = 12;
+                        } elseif($user->rating_id == 7 || $user->rating_id == 11 || $user->rating_id == 12) {
+                            $mdl_rating = 13;
+                        } elseif($user->rating_id == 8 || $user->rating_id == 10) {
+                            $mdl_rating = 14;
+                        } else {
+                            $mdl_rating = 0;
+                        }
 
-                    if($user->rating_id == 1) {
-                        $mdl_rating = 18;
-                    } elseif($user->rating_id == 2) {
-                        $mdl_rating = 9;
-                    } elseif($user->rating_id == 3) {
-                        $mdl_rating = 10;
-                    } elseif($user->rating_id == 4) {
-                        $mdl_rating = 11;
-                    } elseif($user->rating_id == 5) {
-                        $mdl_rating = 12;
-                    } elseif($user->rating_id == 7 || $user->rating_id == 11 || $user->rating_id == 12) {
-                        $mdl_rating = 13;
-                    } elseif($user->rating_id == 8 || $user->rating_id == 10) {
-                        $mdl_rating = 14;
-                    } else {
-                        $mdl_rating = 0;
+                        DB::table('mdl_role_assignments')->insert([
+                            'roleid' => $mdl_rating,
+                            'contextid' => 26,
+                            'userid' => $user->id
+                        ]);
                     }
-
-                    DB::table('mdl_role_assignments')->insert([
-                        'roleid' => $mdl_rating,
-                        'contextid' => 26,
-                        'userid' => $user->id
-                    ]);
                 }
             } else {
                 $visitrej = VisitRej::where('cid', $r->cid)->first();
@@ -391,30 +386,43 @@ class VisitAgreement extends Command
                     }
                     $user->save();
 
-                    //Assigns role in moodle database
-                    if($user->rating_id == 1) {
-                        $mdl_rating = 18;
-                    } elseif($user->rating_id == 2) {
-                        $mdl_rating = 9;
-                    } elseif($user->rating_id == 3) {
-                        $mdl_rating = 10;
-                    } elseif($user->rating_id == 4) {
-                        $mdl_rating = 11;
-                    } elseif($user->rating_id == 5) {
-                        $mdl_rating = 12;
-                    } elseif($user->rating_id == 7 || $user->rating_id == 11 || $user->rating_id == 12) {
-                        $mdl_rating = 13;
-                    } elseif($user->rating_id == 8 || $user->rating_id == 10) {
-                        $mdl_rating = 14;
-                    } else {
-                        $mdl_rating = 0;
-                    }
+                    if(Config::get('app.moodle') == 1) {
+                        //Adds user to moodle database
+                        DB::table('mdl_user')->insert([
+                             'id' => $user->id,
+                             'confirmed' => 1,
+                             'mnethostid' => 1,
+                             'username' => $user->id,
+                             'firstname' => $user->fname,
+                             'lastname' => $user->lname,
+                             'email' => $user->email
+                         ]);
 
-                    DB::table('mdl_role_assignments')->insert([
-                        'roleid' => $mdl_rating,
-                        'contextid' => 26,
-                        'userid' => $user->id
-                    ]);
+                        //Assigns role in moodle database
+                        if($user->rating_id == 1) {
+                            $mdl_rating = 18;
+                        } elseif($user->rating_id == 2) {
+                            $mdl_rating = 9;
+                        } elseif($user->rating_id == 3) {
+                            $mdl_rating = 10;
+                        } elseif($user->rating_id == 4) {
+                            $mdl_rating = 11;
+                        } elseif($user->rating_id == 5) {
+                            $mdl_rating = 12;
+                        } elseif($user->rating_id == 7 || $user->rating_id == 11 || $user->rating_id == 12) {
+                            $mdl_rating = 13;
+                        } elseif($user->rating_id == 8 || $user->rating_id == 10) {
+                            $mdl_rating = 14;
+                        } else {
+                            $mdl_rating = 0;
+                        }
+
+                        DB::table('mdl_role_assignments')->insert([
+                            'roleid' => $mdl_rating,
+                            'contextid' => 26,
+                            'userid' => $user->id
+                        ]);
+                    }
 
                     //Assigns controller initials
                     $user = User::find($r->cid);
@@ -574,10 +582,12 @@ class VisitAgreement extends Command
                     $e->delete();
                 }
 
-                //Sets user as deleted in moodle
-                $moodle = DB::table('mdl_user')->find($use->id);
-                $moodle->deleted = 1;
-                $moodle->save();
+                if(Config::get('app.moodle') == 1) {
+                    //Sets user as deleted in moodle
+                    $moodle = DB::table('mdl_user')->find($use->id);
+                    $moodle->deleted = 1;
+                    $moodle->save();
+                }
 
                 $use->delete();
             }
@@ -598,10 +608,12 @@ class VisitAgreement extends Command
                     $e->delete();
                 }
 
-                //Sets user as deleted in moodle
-                $moodle = DB::table('mdl_user')->find($use->id);
-                $moodle->deleted = 1;
-                $moodle->save();
+                if(Config::get('app.moodle') == 1) {
+                    //Sets user as deleted in moodle
+                    $moodle = DB::table('mdl_user')->find($use->id);
+                    $moodle->deleted = 1;
+                    $moodle->save();
+                }
 
                 $use->delete();
             }
